Estimate maximum possible ctr in group overview
When the "clicks out" is zero the "CTR" is show as zero as well. With
this modification now possible to estimate the upper bound of possible
"CTR". It adds a tooltip which is visible on mouseover on CTR value.

Example:

Let we have 22 clicks and 0 clicks out. If next click would be with
click out through CTR could be 1/23 ~ 4%. So possible CTR lies in range
between 0% and 4%.